+++ scg-impl/src/main/java/com/dwarfeng/scg/impl/service/GenerateServiceImpl.java
@@ -1,0 +1,43 @@
+package com.dwarfeng.scg.impl.service;
+
+import com.dwarfeng.scg.stack.handler.GenerateHandler;
+import com.dwarfeng.scg.stack.service.GenerateService;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import com.dwarfeng.subgrade.stack.log.LogLevel;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class GenerateServiceImpl implements GenerateService {
+
+    private final GenerateHandler generateHandler;
+
+    private final ServiceExceptionMapper sem;
+
+    public GenerateServiceImpl(GenerateHandler generateHandler, ServiceExceptionMapper sem) {
+        this.generateHandler = generateHandler;
+        this.sem = sem;
+    }
+
+    @Override
+    public String generate(StringIdKey scgSettingKey) throws ServiceException {
+        try {
+            return generateHandler.generate(scgSettingKey);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow("基于指定的设置生成序列码时发生异常", LogLevel.WARN, sem, e);
+        }
+    }
+
+    @Override
+    public List<String> batchGenerate(StringIdKey scgSettingKey, int size) throws ServiceException {
+        try {
+            return generateHandler.batchGenerate(scgSettingKey, size);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow("基于指定的设置批量生成序列码时发生异常", LogLevel.WARN, sem, e);
+        }
+    }
+}