+++ scg-impl/src/test/java/com/dwarfeng/scg/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -1,10 +1,7 @@
 package com.dwarfeng.scg.impl.configuration;
 
 import com.dwarfeng.scg.sdk.util.ServiceExceptionCodes;
-import com.dwarfeng.scg.stack.exception.GeneratorException;
-import com.dwarfeng.scg.stack.exception.GeneratorExecutionException;
-import com.dwarfeng.scg.stack.exception.GeneratorMakeException;
-import com.dwarfeng.scg.stack.exception.UnsupportedGeneratorTypeException;
+import com.dwarfeng.scg.stack.exception.*;
 import com.dwarfeng.subgrade.impl.exception.MapServiceExceptionMapper;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
@@ -23,6 +20,7 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(GeneratorExecutionException.class, ServiceExceptionCodes.GENERATOR_EXECUTION_FAILED);
         destination.put(GeneratorMakeException.class, ServiceExceptionCodes.GENERATOR_MAKE_FAILED);
         destination.put(UnsupportedGeneratorTypeException.class, ServiceExceptionCodes.UNSUPPORTED_GENERATOR_TYPE);
+        destination.put(ScgSettingNotExistsException.class, ServiceExceptionCodes.SCG_SETTING_NOT_EXISTS);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINE);
     }
 }